flaw_detection_shivam_sharma/requirements.txt
absl-py==0.8.1
alembic==1.3.1
arrow==0.15.4
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
binaryornot==0.4.4
bleach==3.1.0
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cookiecutter==1.6.0
cycler==0.10.0
cytoolz==0.10.1
dask==2.8.1
decorator==4.4.1
defusedxml==0.6.0
efficientnet-pytorch==0.5.1
entrypoints==0.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Caching==1.7.2
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-SSLify==0.1.5
Flask-Webpack==0.1.0
Flask-WTF==0.14.2
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.16.1
h5py==2.8.0
idna==2.8
imageio==2.6.1
imgaug==0.3.0
importlib-metadata==0.23
imutils==0.5.3
ipykernel==5.1.3
ipyparallel==6.2.4
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jinja2-time==0.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.4
olefile==0.46
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pipenv==2018.11.26
poyo==0.5.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.6.0
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.13.0
SQLAlchemy==1.3.11
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.3.1
tornado==6.0.3
tqdm==4.39.0
traitlets==4.3.3
urllib3==1.25.7
virtualenv==16.7.8
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
